backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
reportlab==4.0.8
openpyxl==3.1.2
pytest==7.4.3
pytest-django==4.7.0
